# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Rent API",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/categories": {
            "get": {
                "tags": [
                    "Categories"
                ],
                "summary": "Get categories",
                "operationId": "categories-list-get",
                "parameters": [
                    {
                        "name": "type",
                        "in": "query",
                        "example": "material"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status_code": {
                                            "type": "integer",
                                            "example": 200
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "name": {
                                                        "type": "string",
                                                        "example": "string"
                                                    },
                                                    "type": {
                                                        "type": "string",
                                                        "enum": [
                                                            "material",
                                                            "product"
                                                        ],
                                                        "example": "product"
                                                    },
                                                    "created_at": {
                                                        "type": "string",
                                                        "example": "2023-06-23 10:10:10"
                                                    },
                                                    "updated_at": {
                                                        "type": "string",
                                                        "example": "2023-06-23 10:10:10"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        },
                                        "page": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "per_page": {
                                            "type": "integer",
                                            "example": 30
                                        },
                                        "total": {
                                            "type": "integer",
                                            "example": 60
                                        },
                                        "last_page": {
                                            "type": "integer",
                                            "example": 2
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Categories"
                ],
                "summary": "Create category",
                "operationId": "category-create",
                "requestBody": {
                    "description": "",
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Category name"
                                    },
                                    "type": {
                                        "type": "string",
                                        "example": "product"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Entity response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status_code": {
                                            "type": "integer",
                                            "example": 200
                                        },
                                        "data": {
                                            "description": "Entity",
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "name": {
                                                    "type": "string",
                                                    "example": "string"
                                                },
                                                "type": {
                                                    "type": "string",
                                                    "enum": [
                                                        "material",
                                                        "product"
                                                    ],
                                                    "example": "product"
                                                },
                                                "created_at": {
                                                    "type": "string",
                                                    "example": "2023-06-23 10:10:10"
                                                },
                                                "updated_at": {
                                                    "type": "string",
                                                    "example": "2023-06-23 10:10:10"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/api/categories/{category}": {
            "get": {
                "tags": [
                    "Categories"
                ],
                "summary": "Get category",
                "operationId": "category-get",
                "parameters": [
                    {
                        "name": "category",
                        "in": "path",
                        "example": 4
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Entity response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status_code": {
                                            "type": "integer",
                                            "example": 200
                                        },
                                        "data": {
                                            "description": "Entity",
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "name": {
                                                    "type": "string",
                                                    "example": "string"
                                                },
                                                "type": {
                                                    "type": "string",
                                                    "enum": [
                                                        "material",
                                                        "product"
                                                    ],
                                                    "example": "product"
                                                },
                                                "created_at": {
                                                    "type": "string",
                                                    "example": "2023-06-23 10:10:10"
                                                },
                                                "updated_at": {
                                                    "type": "string",
                                                    "example": "2023-06-23 10:10:10"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Categories"
                ],
                "summary": "Update category",
                "operationId": "category-update",
                "parameters": [
                    {
                        "name": "category",
                        "in": "path",
                        "example": 4
                    }
                ],
                "requestBody": {
                    "description": "",
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Category name"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Entity response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status_code": {
                                            "type": "integer",
                                            "example": 200
                                        },
                                        "data": {
                                            "description": "Entity",
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "name": {
                                                    "type": "string",
                                                    "example": "string"
                                                },
                                                "type": {
                                                    "type": "string",
                                                    "enum": [
                                                        "material",
                                                        "product"
                                                    ],
                                                    "example": "product"
                                                },
                                                "created_at": {
                                                    "type": "string",
                                                    "example": "2023-06-23 10:10:10"
                                                },
                                                "updated_at": {
                                                    "type": "string",
                                                    "example": "2023-06-23 10:10:10"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Categories"
                ],
                "summary": "Delete category",
                "operationId": "category-delete",
                "parameters": [
                    {
                        "name": "category",
                        "in": "path",
                        "example": 4
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No content response"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearer": {
                "type": "http",
                "name": "bearer",
                "in": "header",
                "scheme": "bearer"
            }
        }
    },
...
}